[MIG] stock_inventory_line_price: Upgrade to v9 * Split models by file * Simplify logic in `stock.inventory.line._resolve_inventory_line` * Remove out of spec manifest keys * Bump version * Update view to use odoo tag & correct view id
