+++ stock_inventory_line_price/README.rst
@@ -10,7 +10,7 @@ Standard price at inventory level
 
 .. image:: https://odoo-community.org/website/image/ir.attachment/5784_f2813bd/datas
    :alt: Try me on Runbot
-   :target: https://runbot.odoo-community.org/runbot/153/8.0
+   :target: https://runbot.odoo-community.org/runbot/153/9.0
 
 Bug Tracker
 ===========
@@ -33,6 +33,7 @@ Contributors
 * Pedro M. Baeza <pedro.baeza@tecnativa.com>
 * Ana Juaristi <ajuaristio@gmail.com>
 * Esther Martín <esthermartin@avanzosc.es>
+* Dave Lasley <dave@laslabs.com>
 
 Maintainer
 ----------
