+++ stock_inventory_line_price/__openerp__.py
@@ -4,17 +4,12 @@
 
 {
     "name": "Standard price at inventory level",
-    "version": "8.0.1.0.0",
+    "version": "9.0.1.0.0",
     "category": "Generic Modules",
     "author": "OdooMRP team, "
               "AvanzOSC, "
               "Tecnativa, "
               "Odoo Community Association (OCA)",
-    "contributors": [
-        "Pedro M. Baeza <pedro.baeza@tecnativa.com>",
-        "Ana Juaristi <ajuaristio@gmail.com>",
-        "Esther Martín <esthermartin@avanzosc.es>",
-    ],
     "website": "http://www.odoomrp.com",
     "depends": [
         "stock",
