+++ stock_inventory_line_price/models/__init__.py
@@ -3,3 +3,4 @@
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
 from . import stock_inventory
+from . import stock_inventory_line
