+++ stock_inventory_line_price/models/stock_inventory.py
@@ -2,33 +2,7 @@
 # (c) 2016 Esther Martín - AvanzOSC
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
-from openerp import api, fields, models
-import openerp.addons.decimal_precision as dp
-
-
-class StockInventoryLine(models.Model):
-    _inherit = 'stock.inventory.line'
-
-    theoretical_std_price = fields.Float(
-        'Theorical Cost Price', digits=dp.get_precision('Product Price'),
-        readonly=True)
-    standard_price = fields.Float('Cost Price',
-                                  digits=dp.get_precision('Product Price'))
-
-    @api.model
-    def _resolve_inventory_line(self, inventory_line):
-        if inventory_line.theoretical_std_price !=\
-                inventory_line.standard_price:
-            inventory_line.product_id.standard_price = \
-                inventory_line.standard_price
-            move_id = super(StockInventoryLine,
-                            self)._resolve_inventory_line(inventory_line)
-            if move_id:
-                move = self.env['stock.move'].browse(move_id)
-                move.price_unit = inventory_line.standard_price
-                return move_id
-        return super(StockInventoryLine, self)._resolve_inventory_line(
-            inventory_line)
+from openerp import api, models
 
 
 class StockInventory(models.Model):
